requirements.txt
django
djangorestframework
south
werkzeug
modestmaps

